♻️ refactoring deploy source a little
📝 dropping some todos for possible refactor improvements for later on
